+++ packages/model-bank/src/aiModels/vertexai.ts
@@ -1,4 +1,4 @@
-import { AIChatModelCard } from '@/types/aiModel';
+import { AIChatModelCard } from '../types/aiModel';
 
 // ref: https://cloud.google.com/vertex-ai/generative-ai/docs/learn/models
 const vertexaiChatModels: AIChatModelCard[] = [
@@ -40,7 +40,6 @@ const vertexaiChatModels: AIChatModelCard[] = [
     description:
       'Gemini 2.5 Pro Preview 是 Google 最先进的思维模型，能够对代码、数学和STEM领域的复杂问题进行推理，以及使用长上下文分析大型数据集、代码库和文档。',
     displayName: 'Gemini 2.5 Pro Preview 05-06',
-    enabled: true,
     id: 'gemini-2.5-pro-preview-05-06',
     maxOutput: 65_536,
     pricing: {
@@ -109,7 +108,6 @@ const vertexaiChatModels: AIChatModelCard[] = [
     contextWindowTokens: 1_048_576 + 65_536,
     description: 'Gemini 2.5 Flash Preview 是 Google 性价比最高的模型，提供全面的功能。',
     displayName: 'Gemini 2.5 Flash Preview 04-17',
-    enabled: true,
     id: 'gemini-2.5-flash-preview-04-17',
     maxOutput: 65_536,
     pricing: {
@@ -128,8 +126,8 @@ const vertexaiChatModels: AIChatModelCard[] = [
     },
     contextWindowTokens: 32_768 + 8192,
     description:
-      'Gemini 2.5 Flash Image Preview 是 Google 最新、最快、最高效的原生多模态模型，它允许您通过对话生成和编辑图像。',
-    displayName: 'Gemini 2.5 Flash Image Preview',
+      'Nano Banana 是 Google 最新、最快、最高效的原生多模态模型，它允许您通过对话生成和编辑图像。',
+    displayName: 'Nano Banana',
     enabled: true,
     id: 'gemini-2.5-flash-image-preview',
     maxOutput: 8192,
@@ -153,7 +151,6 @@ const vertexaiChatModels: AIChatModelCard[] = [
     contextWindowTokens: 1_000_000 + 64_000,
     description: 'Gemini 2.5 Flash-Lite 是 Google 最小、性价比最高的模型，专为大规模使用而设计。',
     displayName: 'Gemini 2.5 Flash-Lite',
-    enabled: true,
     id: 'gemini-2.5-flash-lite',
     maxOutput: 64_000,
     pricing: {
@@ -180,7 +177,6 @@ const vertexaiChatModels: AIChatModelCard[] = [
     description:
       'Gemini 2.5 Flash-Lite Preview 是 Google 最小、性价比最高的模型，专为大规模使用而设计。',
     displayName: 'Gemini 2.5 Flash-Lite Preview 06-17',
-    enabled: true,
     id: 'gemini-2.5-flash-lite-preview-06-17',
     maxOutput: 64_000,
     pricing: {