+++ packages/types/src/llm.ts
@@ -1,10 +1,7 @@
-import { ModelParamsSchema } from 'model-bank';
+import { ModelParamsSchema , AiModelType, Pricing } from 'model-bank';
 import { ReactNode } from 'react';
 
-import { AiModelType, Pricing } from '@/types/aiModel';
-import { AiProviderSettings } from '@/types/aiProvider';
-
-export type ModelPriceCurrency = 'CNY' | 'USD';
+import { AiProviderSettings } from './aiProvider';
 
 export interface ChatModelCard {
   /**
@@ -63,15 +60,6 @@ export interface ChatModelCard {
    */
   vision?: boolean;
 }
-
-export type ResponseAnimationStyle = 'smooth' | 'fadeIn' | 'none';
-export type ResponseAnimation =
-  | {
-      speed?: number;
-      text?: ResponseAnimationStyle;
-      toolsCalling?: ResponseAnimationStyle;
-    }
-  | ResponseAnimationStyle;
 
 export interface ModelProviderCard {
   /**
@@ -153,39 +141,6 @@ export interface ModelProviderCard {
    * provider's website url
    */
   url: string;
-}
-
-// 语言模型的设置参数
-export interface LLMParams {
-  /**
-   * 控制生成文本中的惩罚系数，用于减少重复性
-   * @default 0
-   */
-  frequency_penalty?: number;
-  /**
-   * 生成文本的最大长度
-   */
-  max_tokens?: number;
-  /**
-   * 控制生成文本中的惩罚系数，用于减少主题的变化
-   * @default 0
-   */
-  presence_penalty?: number;
-  /**
-   * 生成文本的随机度量，用于控制文本的创造性和多样性
-   * @default 1
-   */
-  reasoning_effort?: string;
-  /**
-   * 控制模型推理能力
-   * @default medium
-   */
-  temperature?: number;
-  /**
-   * 控制生成文本中最高概率的单个 token
-   * @default 1
-   */
-  top_p?: number;
 }
 
 export type LLMRoleType = 'user' | 'system' | 'assistant' | 'tool';